+++ src/sampler.cpp
@@ -15,7 +15,7 @@ std::random_device Sampler::rd;
 Sampler::Sampler()
 {
     eng = std::ranlux24_base(rd());
-    unif_distr = std::uniform_real_distribution<double>(0, 1);
+    unif_distr = std::uniform_real_distribution<float>(0, 1);
     ber_distr = std::bernoulli_distribution(.5);
 
     for (int i = 0; i < 128; i++)
@@ -24,39 +24,39 @@ Sampler::Sampler()
     }
 }
 
-double Sampler::dbeta(double x, double alpha, double beta, bool return_log)
+float Sampler::dbeta(float x, float alpha, float beta, bool return_log)
 {
     return R::dbeta(x, alpha, beta, return_log);
 }
 
-double Sampler::dpois(int x, double mean, bool return_log)
+float Sampler::dpois(int x, float mean, bool return_log)
 {
     return R::dpois(x, mean, return_log);
 }
 
-double Sampler::dztpois(int x, double lambda)
+float Sampler::dztpois(int x, float lambda)
 {
     return x * std::log(lambda) - std::log(std::exp(lambda) - 1) -
            lgamma_lut[x];
 }
 
-double Sampler::dbinom(int x, int size, double prob, bool return_log)
+float Sampler::dbinom(int x, int size, float prob, bool return_log)
 {
-    double log_p = x * std::log(prob) + (size - x) * std::log(1 - prob);
-    double log_coef = lgamma_lut[size] - lgamma_lut[x] - lgamma_lut[size - x];
+    float log_p = x * std::log(prob) + (size - x) * std::log(1 - prob);
+    float log_coef = lgamma_lut[size] - lgamma_lut[x] - lgamma_lut[size - x];
 
     return return_log ? log_p + log_coef : std::exp(log_p + log_coef);
 }
 
-double Sampler::dgamma(double x, double shape, double scale, bool return_log)
+float Sampler::dgamma(float x, float shape, float scale, bool return_log)
 {
     return R::dgamma(x, shape, scale, return_log);
 }
 
-double Sampler::rgamma(double alpha, double beta)
+float Sampler::rgamma(float alpha, float beta)
 {
-    gamma_distr.param(std::gamma_distribution<double>::param_type(alpha, beta));
-    double x = gamma_distr(eng);
+    gamma_distr.param(std::gamma_distribution<float>::param_type(alpha, beta));
+    float x = gamma_distr(eng);
 
     if (x < UNDERFLO)
     {
@@ -70,24 +70,24 @@ double Sampler::rgamma(double alpha, double beta)
     return x;
 };
 
-double Sampler::rgamma2(double shape, double rate)
+float Sampler::rgamma2(float shape, float rate)
 {
     return R::rgamma(shape, 1 / rate);
 }
 
-std::vector<double> Sampler::rdirichlet(std::vector<double> const &shape_vec)
+std::vector<float> Sampler::rdirichlet(std::vector<float> const &shape_vec)
 {
     int n = shape_vec.size();
-    std::vector<double> res(n);
+    std::vector<float> res(n);
 
-    double res_sum = 0;
+    float res_sum = 0;
     for (int i = 0; i < n; i++)
     {
         res[i] = rgamma(shape_vec[i], 1.0);
         res_sum += res[i];
     }
 
-    double res_sum_inv = 1.0 / res_sum;
+    float res_sum_inv = 1.0 / res_sum;
     for (size_t i = 0; i < res.size(); i++)
     {
         res[i] *= res_sum_inv;
@@ -96,23 +96,23 @@ std::vector<double> Sampler::rdirichlet(std::vector<double> const &shape_vec)
     return res;
 };
 
-std::vector<double> Sampler::rlogit_norm(std::vector<double> const &p,
-                                         double variance)
+std::vector<float> Sampler::rlogit_norm(std::vector<float> const &p,
+                                        float variance)
 {
     int n = p.size() - 1;
 
-    std::vector<double> ret(n + 1);
+    std::vector<float> ret(n + 1);
 
-    double tmp1 = 0;
+    float tmp1 = 0;
     for (int i = 0; i < n; i++)
     {
-        norm_distr.param(std::normal_distribution<double>::param_type(
+        norm_distr.param(std::normal_distribution<float>::param_type(
             log(p[i] / p[n]), variance));
         ret[i] = exp(norm_distr(eng));
         tmp1 += ret[i];
     }
 
-    double tmp2 = 1.0 / (1.0 + tmp1);
+    float tmp2 = 1.0 / (1.0 + tmp1);
     for (int i = 0; i < n; i++)
     {
         ret[i] *= tmp2;
@@ -123,7 +123,7 @@ std::vector<double> Sampler::rlogit_norm(std::vector<double> const &p,
     return ret;
 }
 
-double Sampler::sample_mean_coi(double mean_shape, double mean_scale)
+float Sampler::sample_mean_coi(float mean_shape, float mean_scale)
 {
     return rgamma(mean_shape, mean_scale);
 }
@@ -135,20 +135,20 @@ int Sampler::sample_random_int(int lower, int upper)
     return unif_int_distr(eng);
 }
 
-double Sampler::get_coi_log_prior(int coi, double mean)
+float Sampler::get_coi_log_prior(int coi, float mean)
 {
     return dztpois(coi, mean);
 }
 
-double Sampler::get_coi_mean_log_hyper_prior(double mean, double shape,
-                                             double scale)
+float Sampler::get_coi_mean_log_hyper_prior(float mean, float shape,
+                                            float scale)
 {
     return dgamma(mean, shape, scale, true);
 }
 
 int Sampler::sample_coi_delta() { return (2 * ber_distr(eng) - 1); }
 
-int Sampler::sample_coi_delta(double coi_prop_mean)
+int Sampler::sample_coi_delta(float coi_prop_mean)
 {
     geom_distr.param(std::geometric_distribution<int>::param_type(
         1.0 / (1.0 + coi_prop_mean)));
@@ -156,56 +156,55 @@ int Sampler::sample_coi_delta(double coi_prop_mean)
     return (2 * ber_distr(eng) - 1) * (geom_distr(eng));
 }
 
-double Sampler::get_epsilon_log_prior(double x, double alpha, double beta)
+float Sampler::get_epsilon_log_prior(float x, float alpha, float beta)
 {
     // return dgamma(x, shape, scale, true);
     return dbeta(x, alpha, beta, true);
 }
 
-double Sampler::get_relatedness_log_prior(double x, double alpha, double beta)
+float Sampler::get_relatedness_log_prior(float x, float alpha, float beta)
 {
     return dbeta(x, alpha, beta, true);
 }
 
-double Sampler::sample_epsilon(double curr_epsilon, double variance)
+float Sampler::sample_epsilon(float curr_epsilon, float variance)
 {
     norm_distr.param(
-        std::normal_distribution<double>::param_type(curr_epsilon, variance));
-    double prop = norm_distr(eng);
+        std::normal_distribution<float>::param_type(curr_epsilon, variance));
+    float prop = norm_distr(eng);
     return prop;
 };
 
-std::tuple<double, double> Sampler::sample_constrained(double curr, double var,
-                                                       double lower,
-                                                       double upper)
+std::tuple<float, float> Sampler::sample_constrained(float curr, float var,
+                                                     float lower, float upper)
 {
-    norm_distr.param(std::normal_distribution<double>::param_type(0, var));
-    double eps = norm_distr(eng);
-    double unconstrained = std::log(curr - lower) - std::log(upper - curr);
-    double exp_prop = std::exp(eps + unconstrained);
-    double prop = (upper * exp_prop + lower) / (exp_prop + 1);
+    norm_distr.param(std::normal_distribution<float>::param_type(0, var));
+    float eps = norm_distr(eng);
+    float unconstrained = std::log(curr - lower) - std::log(upper - curr);
+    float exp_prop = std::exp(eps + unconstrained);
+    float prop = (upper * exp_prop + lower) / (exp_prop + 1);
     prop = UtilFunctions::clamp(prop, lower, upper);
 
-    double adj = std::log(prop - lower) + std::log(upper - prop) -
-                 std::log(curr - lower) - std::log(upper - curr);
+    float adj = std::log(prop - lower) + std::log(upper - prop) -
+                std::log(curr - lower) - std::log(upper - curr);
 
     return std::make_tuple(prop, adj);
 }
 
-double Sampler::sample_epsilon_pos(double curr_epsilon_pos, double variance)
+float Sampler::sample_epsilon_pos(float curr_epsilon_pos, float variance)
 {
     return sample_epsilon(curr_epsilon_pos, variance);
 };
 
-double Sampler::sample_epsilon_neg(double curr_epsilon_neg, double variance)
+float Sampler::sample_epsilon_neg(float curr_epsilon_neg, float variance)
 {
     return sample_epsilon(curr_epsilon_neg, variance);
 };
 
-std::vector<double> Sampler::sample_allele_frequencies(
-    std::vector<double> const &curr_allele_frequencies, double alpha)
+std::vector<float> Sampler::sample_allele_frequencies(
+    std::vector<float> const &curr_allele_frequencies, float alpha)
 {
-    std::vector<double> shape_vec(curr_allele_frequencies.size());
+    std::vector<float> shape_vec(curr_allele_frequencies.size());
 
     for (size_t i = 0; i < shape_vec.size(); i++)
     {
@@ -215,8 +214,8 @@ std::vector<double> Sampler::sample_allele_frequencies(
     return rdirichlet(shape_vec);
 };
 
-std::vector<double> Sampler::sample_allele_frequencies2(
-    std::vector<double> const &curr_allele_frequencies, double variance)
+std::vector<float> Sampler::sample_allele_frequencies2(
+    std::vector<float> const &curr_allele_frequencies, float variance)
 {
     return rlogit_norm(curr_allele_frequencies, variance);
 };
@@ -226,16 +225,13 @@ void Sampler::shuffle_vec(std::vector<int> &vec)
     std::shuffle(vec.begin(), vec.end(), eng);
 }
 
-double Sampler::sample_unif() { return unif_distr(eng); };
+float Sampler::sample_unif() { return unif_distr(eng); };
 
-double Sampler::sample_log_mh_acceptance()
-{
-    return std::log(unif_distr(eng));
-};
+float Sampler::sample_log_mh_acceptance() { return std::log(unif_distr(eng)); };
 
 LatentGenotype Sampler::sample_latent_genotype(
-    const std::vector<int> &obs_genotype, int coi, double epsilon_pos,
-    double epsilon_neg)
+    const std::vector<int> &obs_genotype, int coi, float epsilon_pos,
+    float epsilon_neg)
 {
     int total_alleles = obs_genotype.size();
     int total_obs_positives = 0;
@@ -272,8 +268,8 @@ LatentGenotype Sampler::sample_latent_genotype(
     }
     int total_true_negatives = total_obs_negatives - total_false_negatives;
 
-    double log_prob_total_false_negatives =
-        std::log(boost::math::binomial_coefficient<double>(
+    float log_prob_total_false_negatives =
+        std::log(boost::math::binomial_coefficient<float>(
             total_obs_negatives, total_false_negatives - min_false_negatives)) +
         (total_false_negatives - min_false_negatives) *
             std::log(epsilon_neg / obs_genotype.size()) +
@@ -296,8 +292,8 @@ LatentGenotype Sampler::sample_latent_genotype(
     }
     int total_true_positives = total_obs_positives - total_false_positives;
 
-    double log_prob_total_false_positives =
-        std::log(boost::math::binomial_coefficient<double>(
+    float log_prob_total_false_positives =
+        std::log(boost::math::binomial_coefficient<float>(
             total_obs_positives, total_false_positives - min_false_positives)) +
         (total_false_positives - min_false_positives) *
             std::log(epsilon_pos / obs_genotype.size()) +
@@ -321,16 +317,16 @@ LatentGenotype Sampler::sample_latent_genotype(
     assert(allele_index_vec.size() ==
            total_true_positives + total_false_negatives);
 
-    double log_prob_positive_indices =
-        -std::log(boost::math::binomial_coefficient<double>(
+    float log_prob_positive_indices =
+        -std::log(boost::math::binomial_coefficient<float>(
             total_obs_positives, total_true_positives));
-    double log_prob_negative_indices =
-        -std::log(boost::math::binomial_coefficient<double>(
+    float log_prob_negative_indices =
+        -std::log(boost::math::binomial_coefficient<float>(
             total_obs_negatives, total_false_negatives));
 
-    double log_prob = log_prob_positive_indices + log_prob_negative_indices +
-                      log_prob_total_false_positives +
-                      log_prob_total_false_negatives;
+    float log_prob = log_prob_positive_indices + log_prob_negative_indices +
+                     log_prob_total_false_positives +
+                     log_prob_total_false_negatives;
 
     return LatentGenotype{allele_index_vec, log_prob};
 }
